chore: ignore shm proxy in coverage
