+++ .ci/run_tests.sh
@@ -5,7 +5,7 @@ set -e
 if [ "$OPENRESTY_TESTS" != "yes" ]; then
   make lint
   busted -v --coverage -o gtest --repeat 1
-  luacov-coveralls -i src/cassandra -e bit.lua -e socket.lua
+  luacov-coveralls -i src/cassandra -e bit.lua -e socket.lua -e shm.lua
 else
   make prove
 fi
