+++ lua-cassandra-0.5.0-0.rockspec
@@ -35,6 +35,7 @@ build = {
 
     ["cassandra.auth.plain_text_provider"] = "src/cassandra/auth/plain_text_provider.lua",
 
+    ["cassandra.utils.shm"] = "src/cassandra/utils/shm.lua",
     ["cassandra.utils.bit"] = "src/cassandra/utils/bit.lua",
     ["cassandra.utils.time"] = "src/cassandra/utils/time.lua",
     ["cassandra.utils.table"] = "src/cassandra/utils/table.lua",
