+++ src/cassandra/cache.lua
@@ -1,120 +1,22 @@
 local log = require "cassandra.log"
 local Errors = require "cassandra.errors"
 local string_utils = require "cassandra.utils.string"
-
-local table_concat = table.concat
+local concat = table.concat
 local tonumber = tonumber
 
-local in_ngx = ngx ~= nil
 local dicts = {}
-
--- ngx.shared.DICT proxy
--- https://github.com/bsm/fakengx/blob/master/fakengx.lua
--- Used when the driver is required outside of ngx_lua.
--- Eventually, attaching the cluster infos to the session
--- should be done standardly, but this is a faster alternative
--- for now.
-
-local SharedDict = {}
-
-local function set(data, key, value)
-  data[key] = {
-    value = value,
-    info = {expired = false}
-  }
-end
-
-function SharedDict:new()
-  return setmetatable({data = {}}, {__index = self})
-end
-
-function SharedDict:get(key)
-  return self.data[key] and self.data[key].value, nil
-end
-
-function SharedDict:set(key, value)
-  set(self.data, key, value)
-  return true, nil, false
-end
-
-function SharedDict:get_keys()
-  local keys = {}
-  for k in pairs(self.data) do
-    keys[#keys + 1] = k
-  end
-  return keys
-end
-
-SharedDict.safe_set = SharedDict.set
-
-function SharedDict:add(key, value)
-  if self.data[key] ~= nil then
-    return false, "exists", false
-  end
-
-  set(self.data, key, value)
-  return true, nil, false
-end
-
-function SharedDict:replace(key, value)
-  if self.data[key] == nil then
-    return false, "not found", false
-  end
-
-  set(self.data, key, value)
-  return true, nil, false
-end
-
-function SharedDict:delete(key)
-  self.data[key] = nil
-end
-
-function SharedDict:incr(key, value)
-  if not self.data[key] then
-    return nil, "not found"
-  elseif type(self.data[key].value) ~= "number" then
-    return nil, "not a number"
-  end
-
-  self.data[key].value = self.data[key].value + value
-  return self.data[key].value, nil
-end
-
-function SharedDict:flush_all()
-  for _, item in pairs(self.data) do
-    item.info.expired = true
-  end
-end
-
-function SharedDict:flush_expired(n)
-  local data = self.data
-  local flushed = 0
-
-  for key, item in pairs(self.data) do
-    if item.info.expired then
-      data[key] = nil
-      flushed = flushed + 1
-      if n and flushed == n then
-        break
-      end
-    end
-  end
-
-  self.data = data
-
-  return flushed
-end
 
 local function get_dict(shm)
   local dict = dicts[shm]
 
   if dict == nil then
-    if in_ngx then
+    if ngx ~= nil then
       dict = ngx.shared[shm]
       if dict == nil then
         error("No shared dict named "..shm)
       end
     else
+      local SharedDict = require "cassandra.utils.shm"
       dict = SharedDict:new()
     end
     dicts[shm] = dict
@@ -131,7 +33,7 @@ local _SEP = ";"
 
 local function set_hosts(shm, hosts)
   local dict = get_dict(shm)
-  local ok, err = dict:safe_set(_HOSTS_KEY, table_concat(hosts, _SEP))
+  local ok, err = dict:safe_set(_HOSTS_KEY, concat(hosts, _SEP))
   if not ok then
     return false, Errors.shm(shm, "cannot store hosts ("..err..")")
   end
